Refactor VideoExplorerPage.jsx to reduce lines
