+++ frontend_react/src/components/VideoExplorer/index.ts
@@ -1,0 +1,2 @@
+export { default as VideoTableHeader } from "./VideoTableHeader";
+export { default as MobileSortDropdown } from "./MobileSortDropdown";
