+++ frontend_react/src/components/common/index.ts
@@ -1,0 +1,4 @@
+export { default as ChannelInfoCard } from "./ChannelInfoCard";
+export { default as TopLevelTabs } from "./TopLevelTabs";
+export { default as SubCategoryTabs } from "./SubCategoryTabs";
+export { default as VideoCard } from "./VideoCard";
