+++ frontend_react/src/pages/VideoExplorerPage.jsx
@@ -2,19 +2,22 @@ import React, { useEffect, useState, useMemo } from "react";
 import { useSearchParams } from "react-router-dom";
 import { toast } from "react-hot-toast";
 
-import { useClassifiedVideos } from "../hooks/useClassifiedVideos";
-import { useVideoBrowseState } from "../hooks/useVideoBrowseState";
-import { useChartControlState } from "../hooks/useChartControlState";
-import useVideoSortControl from "../hooks/useVideoSortControl";
+import {
+  useClassifiedVideos, useVideoBrowseState, useChartControlState,
+  useAutoUpdateVideos, useVideoSortControl,
+} from "../hooks";
 
-import ChannelInfoCard from "../components/common/ChannelInfoCard";
-import TopLevelTabs from "../components/common/TopLevelTabs";
-import SubCategoryTabs from "../components/common/SubCategoryTabs";
+import {
+  ChannelInfoCard, TopLevelTabs, SubCategoryTabs, VideoCard
+} from "../components/common";
+
+import {
+  VideoTableHeader, MobileSortDropdown,
+} from "../components/VideoExplorer";
+
+import { sortVideos } from "../utils/sortVideos";
 import CategoryChartSection from "../components/chart/CategoryChartSection";
-import VideoCard from "../components/common/VideoCard"
 import MainLayout from "../components/layout/MainLayout";
-import VideoTableHeader from "../components/VideoExplorer/VideoTableHeader";
-import MobileSortDropdown from "../components/VideoExplorer/MobileSortDropdown";
 
 // ✅ 若 URL 無指定 channel，使用預設頻道
 const ADMIN_CHANNEL_ID = "UCLxa0YOtqi8IR5r2dSLXPng";
@@ -27,17 +30,17 @@ const VideoExplorerPage = () => {
     handleSortChange,
   } = useVideoSortControl("publishDate");
 
-  /* ---------------- 1. 解析 URL 參數 ---------------- */
+  /* ---------------- 解析 URL 參數 ---------------- */
   const [searchParams] = useSearchParams();
   const channelId = searchParams.get("channel") || ADMIN_CHANNEL_ID;
 
-  /* ---------------- 2. 讀取影片與分類快取 ---------------- */
+  /* ---------------- 讀取影片與分類快取 ---------------- */
   const { videos, loading, error } = useClassifiedVideos(
     channelId,
     "videos"
   );
 
-  /* ---------------- 3. 處理瀏覽狀態 ---------------- */
+  /* ---------------- 處理瀏覽狀態 ---------------- */
   const {
     SORT_FIELDS,
     videoType,
@@ -47,22 +50,12 @@ const VideoExplorerPage = () => {
     filteredVideos,
   } = useVideoBrowseState(videos);
 
-  const sortedVideos = useMemo(() => {
-    const list = [...filteredVideos];
+  const sortedVideos = useMemo(
+    () => sortVideos(filteredVideos, sortField, sortOrder),
+    [filteredVideos, sortField, sortOrder]
+  );
 
-    list.sort((a, b) => {
-      const aValue = a[sortField] || "";
-      const bValue = b[sortField] || "";
-
-      if (aValue < bValue) return sortOrder === "asc" ? -1 : 1;
-      if (aValue > bValue) return sortOrder === "asc" ? 1 : -1;
-      return 0;
-    });
-
-    return list;
-  }, [filteredVideos, sortField, sortOrder]);
-
-  /* ---------------- 4. 圖表控制 ---------------- */
+  /* ---------------- 圖表控制 ---------------- */
   const {
     chartType,
     setChartType,
@@ -70,13 +63,12 @@ const VideoExplorerPage = () => {
     setDurationUnit,
   } = useChartControlState();
 
-  /* ---------------- 5. 切換頻道完成後關閉 Toast ---------------- */
+  /* ---------------- 切換頻道完成後關閉 Toast ---------------- */
   useEffect(() => {
     if (!loading) toast.dismiss("channel-switch");
   }, [loading]);
 
-  const [drawerOpen, setDrawerOpen] = useState(false);
-  const [sidebarCollapsed, setSidebarCollapsed] = useState(false);
+  const [setDrawerOpen] = useState(false);
 
   useEffect(() => {
     const handler = () => setDrawerOpen(true);
@@ -84,39 +76,9 @@ const VideoExplorerPage = () => {
     return () => window.removeEventListener("open-channel-drawer", handler);
   }, []);
 
-  useEffect(() => {
-    const runUpdateCheck = async () => {
-      try {
-        const res = await fetch(
-          `${import.meta.env.VITE_API_BASE}/api/videos/check-update?channelId=${channelId}`
-        );
-        const data = await res.json();
+  useAutoUpdateVideos(channelId);
 
-        if (data.shouldUpdate && data.updateToken) {
-          await fetch(`${import.meta.env.VITE_API_BASE}/api/videos/update`, {
-            method: "POST",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({
-              channelId,
-              updateToken: data.updateToken,
-            }),
-          });
-        }
-      } catch (e) {
-        console.warn("🔁 頻道初始化自動更新失敗", e);
-      }
-    };
-
-    runUpdateCheck();
-  }, [channelId]);
-
-  /* ---------------- 6. 排序箭頭 ---------------- */
-  const arrowOf = (field) => {
-    if (field !== sortField) return null;
-    return sortOrder === "asc" ? "🔼" : "🔽";
-  };
-
-  /* ---------------- 7. 主要畫面 ---------------- */
+  /* ---------------- 主要畫面 ---------------- */
   return (
     <MainLayout>
       <ChannelInfoCard />
