+++ frontend_react/src/types/video.ts
@@ -1,0 +1,4 @@
+export type VideoItem = {
+  videoId: string;
+  [key: string]: any; // 允許其他動態欄位，如 publishDate, duration 等
+};
